py/parse: Fix corner cases broken by support for float const folding
The float const folding code should not fold expressions resulting
into complex results, as the mpy parser for complex immediates
has limitations.

The constant generation code must distinguish between -0.0 and 0.0,
which are different even if C consider them as ==

The constant generation code should use a specific representation
for floating point constants, that is bijective with regard to
mp_parse_num_float(). As that function exhibit some unusual quirks
due to mantissa overflow when parsing numbers with many decimals,
the only way to know if an extra digit should be added is to call
mp_parse_num_float() at compile time on the generated string.

Signed-off-by: Yoctopuce dev <dev@yoctopuce.com>