py/obj: Prevent undesired switch from floats to obj in REPR_C.
The current implementation of mp_obj_new_float(f) in REPR_C
may create non-float objects depending on the value of f.
This can happen in particular with handcrafted float values
via `array.array` or `struct`. Any nan-like float following
the pattern `01111111 1xxxxxxx xxxxxxxx xxxxx1xx` would be
switched to an immediate object or a qstr string. When
the qstr index is too large, this does cause a crash.

This commit fixes the issue, and adds the relevant test cases.

The extra check would have increased the overall code size,
but we are compensating for it by removing an unneeded
bit clearing in the opposite function the retrieves a float
value. Indeed, leaving bit 1 set provides a better guess of
the original float than biasing the result by always truncating
toward 0.

Signed-off-by: Yoctopuce dev <dev@yoctopuce.com>