py/objint_longlong: Add interoperability with floats.
Current longlong code does not support mixed operation with
floating point numbers. This commit applies the same method
used in mpz functions to allow interoperability with floats.

To make sure this code is tested, the longlong build is
updated to include floating point support, using REPR_C
as it is a typical setup for limited bare metal ports.

Signed-off-by: Yoctopuce dev <dev@yoctopuce.com>